refactor: clean architecture
