+++ app/components/Card/SocialCard.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { SocialCardProps } from '../@types/social.d';
-import styles from '../styles/SocialCard.module.scss';
+import { SocialCardProps } from '../../@types/social';
+import styles from './SocialCard.module.scss';
 
 export const SocialCard: React.FC<SocialCardProps> = ({ platform, link }) => {
   return (
