+++ app/components/Layout/Footer.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import styles from '../styles/Footer.module.scss';
+import styles from './Footer.module.scss';
 
 export const Footer: React.FC = () => {
   const year = new Date().getFullYear();
