--- src/app/layout.tsx
@@ -1,20 +1,0 @@
-'use client';
-import type { Metadata } from 'next';
-import './globals.css';
-
-const metadata: Metadata = {
-  title: 'Thiesset Fabien - Portfolio',
-  description: 'My personal portfolio crafted with ❤️ using Next.js',
-};
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang='fr'>
-      <body>{children}</body>
-    </html>
-  );
-}
