--- src/app/styles/Navbar.module.scss
@@ -1,36 +1,0 @@
-.navbar {
-  height: 14.2rem;
-  width: 100vw;
-  background-color: green;
-  padding: 3.5vw 3rem;
-  justify-content: center;
-  display: flex;
-
-  &__content {
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    width: 100%;
-    max-width: 1200px;
-    opacity: 1;
-    filter: blur(0);
-    transform: none;
-    background-color: hsla(0, 0%, 100%, 0.1);
-    padding: 0.1rem;
-    border-radius: 2.4rem;
-    position: relative;
-    gap: 0.25rem;
-  }
-
-  &__filter {
-    background-image: radial-gradient(
-      circle at 50% 150%,
-      rgba(32, 141, 229, 0.17),
-      transparent
-    );
-    display: flex;
-    align-items: center;
-    height: 3.2rem;
-    padding: 0 3.2rem;
-  }
-}
